1. Add general error when getting ApiException in the while doing the restorable check 2. Move the API call to AuthApi class. 3. Add another step(and string) to the progress bar for this case. 4. Fix namings
